+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.2.0 IN PROGRESS
   * SI-172 Add Duplicate of Sequences for number generator sequences in settings
+  * SI-151 Enable dynamic ${current_year} token in number generator output templates
 
 ## 4.1.1 2026-05-28
   * Updated dependencies for Trillium release